7/20

+++ api/dbops/dbops.go
@@ -3,7 +3,6 @@ package dbops
 import (
 	"database/sql"
 	"log"
-	"time"
 	"video-streaming/api/defs"
 
 	_ "github.com/lib/pq"
@@ -69,14 +68,12 @@ func GetUser(user_name string) (*defs.User, error) { //取得使用者資料
 	return res, nil
 }
 
-func AddNewVideo(author_name string, title string) error {
-	t := time.Now()
-	ctime := t.Format("2006-01-02") // YY-MM-DD
-	stmtIns, err := db.Prepare("INSERT INTO video_info (author_name, video_title, create_time, viewed) VALUES($1, $2, $3, 0)")
+func AddNewVideo(author_name string, title string) error { //新增影片
+	stmtIns, err := db.Prepare("INSERT INTO video_info (author_name, video_title, create_time, viewed) VALUES($1, $2, NOW(), 0)")
 	if err != nil {
 		return err
 	}
-	_, err = stmtIns.Exec(author_name, title, ctime)
+	_, err = stmtIns.Exec(author_name, title)
 	if err != nil {
 		return err
 	}
@@ -84,15 +81,46 @@ func AddNewVideo(author_name string, title string) error {
 	return nil
 }
 
-func ListVideoInfo() ([]*defs.VideoInfo, error) {
-	stmtOut, err := db.Prepare("SELECT * FROM video_info ORDER BY create_time DESC")
-	var res []*defs.VideoInfo
+func GetVideoInfo(vid int) (*defs.VideoInfo, error) {
+	stmtOut, err := db.Prepare("SELECT author_name, video_title, create_time, viewed FROM video_info WHERE video_id=$1")
 	if err != nil {
-		return res, err
+		return nil, err
 	}
-	rows, err := stmtOut.Query()
-	if err != nil {
-		return res, err
+	var author, title, ctime string
+	var view int
+	err = stmtOut.QueryRow(vid).Scan(&author, &title, &ctime, &view)
+	res := &defs.VideoInfo{
+		Video_id:    vid,
+		Author_name: author,
+		Video_title: title,
+		Create_time: ctime,
+		Viewed:      view,
+	}
+	return res, nil
+}
+
+func ListVideoInfo(username string) ([]*defs.VideoInfo, error) { //顯示影片
+	var res []*defs.VideoInfo
+	var rows *sql.Rows
+	var stmtOut *sql.Stmt
+	if username == "" { //所有
+		stmtOut, err = db.Prepare("SELECT * FROM video_info ORDER BY create_time DESC")
+		if err != nil {
+			return res, err
+		}
+		rows, err = stmtOut.Query()
+		if err != nil {
+			return res, err
+		}
+	} else { //僅使用者上傳
+		stmtOut, err = db.Prepare("SELECT * FROM video_info WHERE author_name=$1 ORDER BY create_time DESC")
+		if err != nil {
+			return res, err
+		}
+		rows, err = stmtOut.Query(username)
+		if err != nil {
+			return res, err
+		}
 	}
 	for rows.Next() {
 		var author, title, ctime string
@@ -114,12 +142,64 @@ func ListVideoInfo() ([]*defs.VideoInfo, error) {
 	return res, nil
 }
 
-func DeleteVideoInfo(vid int) error {
-	stmtDel, err := db.Prepare("DELETE FROM video_info WHERE video_id = $1")
+func DeleteVideoInfo(vid int,uname string) error { //刪除影片
+	stmtDel, err := db.Prepare("DELETE FROM video_info WHERE video_id = $1 and author_name = $2")
 	if err != nil {
 		return err
 	}
-	_, err = stmtDel.Exec(vid)
+	_, err = stmtDel.Exec(vid,uname)
+	if err != nil {
+		return err
+	}
+	defer stmtDel.Close()
+	return nil
+}
+
+func AddNewComments(vid int, user_name string, content string) error { //新增評論
+	stmtIns, err := db.Prepare("INSERT INTO comments (user_name, video_id, contents, record_time) VALUES ($1, $2, $3, NOW())")
+	if err != nil {
+		return err
+	}
+	_, err = stmtIns.Exec(user_name, vid, content)
+	if err != nil {
+		return err
+	}
+	defer stmtIns.Close()
+	return nil
+}
+
+func ListComments(vid int) ([]*defs.Comment, error) { //顯示評論
+	stmtOut, err := db.Prepare("SELECT * from comments where video_id=$1")
+	var res []*defs.Comment
+	rows, err := stmtOut.Query(vid)
+	if err != nil {
+		return res, err
+	}
+	for rows.Next() {
+		var name, content, time string
+		var id int
+		if err := rows.Scan(&id, &name, &content, &vid, &time); err != nil {
+			return res, err
+		}
+		c := &defs.Comment{
+			Comment_id:  id,
+			User_name:   name,
+			Content:     content,
+			Video_id:    vid,
+			Record_time: time[:19],
+		}
+		res = append(res, c)
+	}
+	defer stmtOut.Close()
+	return res, nil
+}
+
+func DeleteCommentInfo(cid int,uname string) error{		//刪除評論
+	stmtDel, err := db.Prepare("DELETE FROM comments WHERE comment_id = $1 and user_name = $2")
+	if err != nil {
+		return err
+	}
+	_, err = stmtDel.Exec(cid,uname)
 	if err != nil {
 		return err
 	}
